fix bug input form

+++ database/migrations/2025_01_23_063247_halaman__donasi.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('donasi', function (Blueprint $table) {
             $table->id();
             $table->string('program');
-            $table->integer('nominal');
+            $table->unsignedBigInteger('nominal');
             $table->string('foto_bukti')->nullable();
             $table->string('nama_donatur');
             $table->string('alamat');
